Extract `ErrorResponses` from application controller
